♻️ refactor: refactor the globalStore usage
