+++ src/app/chat/(desktop)/features/ChatHeader/HeaderAction.tsx
@@ -4,8 +4,8 @@ import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { DESKTOP_HEADER_ICON_SIZE } from '@/const/layoutTokens';
+import { useGlobalStore } from '@/store/global';
 import { featureFlagsSelectors, useServerConfigStore } from '@/store/serverConfig';
-import { useUserStore } from '@/store/user';
 
 import SettingButton from '../../../features/SettingButton';
 import ShareButton from '../../../features/ShareButton';
@@ -13,7 +13,7 @@ import ShareButton from '../../../features/ShareButton';
 const HeaderAction = memo(() => {
   const { t } = useTranslation('chat');
 
-  const [showAgentSettings, toggleConfig] = useUserStore((s) => [
+  const [showAgentSettings, toggleConfig] = useGlobalStore((s) => [
     s.preference.showChatSideBar,
     s.toggleChatSideBar,
   ]);
