+++ src/chains/__tests__/summaryDescription.test.ts
@@ -5,7 +5,7 @@ import { globalHelpers } from '@/store/user/helpers';
 import { chainSummaryDescription } from '../summaryDescription';
 
 // Mock the globalHelpers.getCurrentLanguage function
-vi.mock('@/store/global/helpers', () => ({
+vi.mock('@/store/user/helpers', () => ({
   globalHelpers: {
     getCurrentLanguage: vi.fn(() => 'en-US'),
   },
