--- src/store/global/helpers.ts
@@ -1,8 +1,0 @@
-import { settingsSelectors } from './slices/settings/selectors';
-import { useGlobalStore } from './store';
-
-const getCurrentLanguage = () => settingsSelectors.currentLanguage(useGlobalStore.getState());
-
-export const globalHelpers = {
-  getCurrentLanguage,
-};
