+++ src/store/global/selectors.ts
@@ -1,8 +1,9 @@
-export { commonSelectors } from './slices/common/selectors';
-export { preferenceSelectors } from './slices/preference/selectors';
-export {
-  modelConfigSelectors,
-  modelProviderSelectors,
-  settingsSelectors,
-  syncSettingsSelectors,
-} from './slices/settings/selectors';
+import { GlobalStore } from '@/store/global';
+import { SessionDefaultGroup } from '@/types/session';
+
+const sessionGroupKeys = (s: GlobalStore): string[] =>
+  s.preference.expandSessionGroupKeys || [SessionDefaultGroup.Pinned, SessionDefaultGroup.Default];
+
+export const preferenceSelectors = {
+  sessionGroupKeys,
+};
