--- src/store/global/slices/common/selectors.ts
@@ -1,6 +1,0 @@
-import { GlobalStore } from '@/store/global';
-
-export const commonSelectors = {
-  userAvatar: (s: GlobalStore) => s.avatar || '',
-  userId: (s: GlobalStore) => s.userId,
-};
