--- src/store/global/slices/settings/actions/general.ts
@@ -1,56 +1,0 @@
-import { ThemeMode } from 'antd-style';
-import isEqual from 'fast-deep-equal';
-import { DeepPartial } from 'utility-types';
-import type { StateCreator } from 'zustand/vanilla';
-
-import { userService } from '@/services/user';
-import type { GlobalStore } from '@/store/global';
-import { LobeAgentSettings } from '@/types/session';
-import { GlobalSettings } from '@/types/settings';
-import { difference } from '@/utils/difference';
-import { merge } from '@/utils/merge';
-
-export interface GeneralSettingsAction {
-  importAppSettings: (settings: GlobalSettings) => Promise<void>;
-  resetSettings: () => Promise<void>;
-  setSettings: (settings: DeepPartial<GlobalSettings>) => Promise<void>;
-  switchThemeMode: (themeMode: ThemeMode) => Promise<void>;
-  updateDefaultAgent: (agent: DeepPartial<LobeAgentSettings>) => Promise<void>;
-}
-
-export const generalSettingsSlice: StateCreator<
-  GlobalStore,
-  [['zustand/devtools', never]],
-  [],
-  GeneralSettingsAction
-> = (set, get) => ({
-  importAppSettings: async (importAppSettings) => {
-    const { setSettings } = get();
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { password: _, ...settings } = importAppSettings;
-
-    await setSettings(settings);
-  },
-  resetSettings: async () => {
-    await userService.resetUserSettings();
-    await get().refreshUserConfig();
-  },
-  setSettings: async (settings) => {
-    const { settings: prevSetting, defaultSettings } = get();
-
-    const nextSettings = merge(prevSetting, settings);
-
-    if (isEqual(prevSetting, nextSettings)) return;
-
-    const diffs = difference(nextSettings, defaultSettings);
-
-    await userService.updateUserSettings(diffs);
-    await get().refreshUserConfig();
-  },
-  switchThemeMode: async (themeMode) => {
-    await get().setSettings({ themeMode });
-  },
-  updateDefaultAgent: async (defaultAgent) => {
-    await get().setSettings({ defaultAgent });
-  },
-});
