--- src/store/global/slices/settings/selectors/index.ts
@@ -1,4 +1,0 @@
-export { modelConfigSelectors } from './modelConfig';
-export { modelProviderSelectors } from './modelProvider';
-export { settingsSelectors } from './settings';
-export { syncSettingsSelectors } from './sync';
