+++ src/store/global/store.ts
@@ -5,20 +5,16 @@ import { StateCreator } from 'zustand/vanilla';
 
 import { isDev } from '@/utils/env';
 
+import { type GlobalStoreAction, globalActionSlice } from './action';
 import { type GlobalState, initialState } from './initialState';
-import { type CommonAction, createCommonSlice } from './slices/common/action';
-import { type PreferenceAction, createPreferenceSlice } from './slices/preference/action';
-import { type SettingsAction, createSettingsSlice } from './slices/settings/actions';
 
 //  ===============  聚合 createStoreFn ============ //
 
-export type GlobalStore = CommonAction & GlobalState & SettingsAction & PreferenceAction;
+export type GlobalStore = GlobalState & GlobalStoreAction;
 
 const createStore: StateCreator<GlobalStore, [['zustand/devtools', never]]> = (...parameters) => ({
   ...initialState,
-  ...createCommonSlice(...parameters),
-  ...createSettingsSlice(...parameters),
-  ...createPreferenceSlice(...parameters),
+  ...globalActionSlice(...parameters),
 });
 
 //  ===============  实装 useStore ============ //
