+++ src/store/user/slices/preference/action.ts
@@ -1,4 +1,3 @@
-import { produce } from 'immer';
 import { SWRResponse } from 'swr';
 import type { StateCreator } from 'zustand/vanilla';
 
@@ -11,14 +10,7 @@ import type { GlobalPreference, Guide } from './initialState';
 
 const n = setNamespace('preference');
 
-/**
- * 设置操作
- */
 export interface PreferenceAction {
-  toggleChatSideBar: (visible?: boolean) => void;
-  toggleExpandSessionGroup: (id: string, expand: boolean) => void;
-  toggleMobileTopic: (visible?: boolean) => void;
-  toggleSystemRole: (visible?: boolean) => void;
   updateGuideState: (guide: Partial<Guide>) => void;
   updatePreference: (preference: Partial<GlobalPreference>, action?: any) => void;
   useInitPreference: () => SWRResponse;
@@ -30,37 +22,6 @@ export const createPreferenceSlice: StateCreator<
   [],
   PreferenceAction
 > = (set, get) => ({
-  toggleChatSideBar: (newValue) => {
-    const showChatSideBar =
-      typeof newValue === 'boolean' ? newValue : !get().preference.showChatSideBar;
-
-    get().updatePreference({ showChatSideBar }, n('toggleAgentPanel', newValue));
-  },
-  toggleExpandSessionGroup: (id, expand) => {
-    const { preference } = get();
-    const nextExpandSessionGroup = produce(preference.expandSessionGroupKeys, (draft) => {
-      if (expand) {
-        if (draft.includes(id)) return;
-        draft.push(id);
-      } else {
-        const index = draft.indexOf(id);
-        if (index !== -1) draft.splice(index, 1);
-      }
-    });
-    get().updatePreference({ expandSessionGroupKeys: nextExpandSessionGroup });
-  },
-  toggleMobileTopic: (newValue) => {
-    const mobileShowTopic =
-      typeof newValue === 'boolean' ? newValue : !get().preference.mobileShowTopic;
-
-    get().updatePreference({ mobileShowTopic }, n('toggleMobileTopic', newValue));
-  },
-  toggleSystemRole: (newValue) => {
-    const showSystemRole =
-      typeof newValue === 'boolean' ? newValue : !get().preference.mobileShowTopic;
-
-    get().updatePreference({ showSystemRole }, n('toggleMobileTopic', newValue));
-  },
   updateGuideState: (guide) => {
     const { updatePreference } = get();
     const nextGuide = merge(get().preference.guide, guide);
