+++ src/store/user/slices/preference/selectors.ts
@@ -1,8 +1,4 @@
 import { UserStore } from '@/store/user';
-import { SessionDefaultGroup } from '@/types/session';
-
-const sessionGroupKeys = (s: UserStore): string[] =>
-  s.preference.expandSessionGroupKeys || [SessionDefaultGroup.Pinned, SessionDefaultGroup.Default];
 
 const useCmdEnterToSend = (s: UserStore): boolean => s.preference.useCmdEnterToSend || false;
 
@@ -12,7 +8,6 @@ const hideSyncAlert = (s: UserStore) => s.preference.hideSyncAlert;
 
 export const preferenceSelectors = {
   hideSyncAlert,
-  sessionGroupKeys,
   useCmdEnterToSend,
   userAllowTrace,
 };
